% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diann_cleanup_test.R
\name{diann.cleanup.new}
\alias{diann.cleanup.new}
\title{This function cleans up and analyzes the DIA output from DIANN, adding a pseudocount}
\usage{
diann.cleanup.new(df, kinases, metadata, directory, unique_df, peptide)
}
\arguments{
\item{df}{Unique protein groups matrix from DIANN}

\item{kinases}{Either the human or mouse kinome spreadsheet}

\item{metadata}{File with column titled "Sample.ID" with the column names of the abundance values in the protein groups file, and a column titled "Treatment" with the corresponding treatment replicate}

\item{directory}{Output folder}

\item{unique_df}{Either "protein" for pg.matrix or "gene" for unique_gene.matrix}

\item{peptides}{report.parquet from DIANN}
}
\description{
This function cleans up and analyzes the DIA output from DIANN, adding a pseudocount
}
\examples{
diann.cleanup.new()
}
\keyword{DIANN}
