% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diann.cleanup.test.R
\name{diann.cleanup.test}
\alias{diann.cleanup.test}
\title{This function cleans up and analyzes the DIA output from DIANN, adding a pseudocount and imputing}
\usage{
diann.cleanup.test(
  df,
  kinases,
  metadata,
  directory,
  unique_df,
  peptide,
  impute_seed = 123
)
}
\arguments{
\item{df}{Unique protein groups matrix from DIANN}

\item{kinases}{Either the human or mouse kinome spreadsheet}

\item{metadata}{File with column titled "Sample.ID" with the column names of the abundance values in the protein groups file, and a column titled "Treatment" with the corresponding treatment replicate}

\item{directory}{Output folder}

\item{unique_df}{Either "protein" for pg.matrix or "gene" for unique_gene.matrix}

\item{peptides}{report.pr.matrix from DIANN}
}
\description{
This function cleans up and analyzes the DIA output from DIANN, adding a pseudocount and imputing
}
\examples{
diann.cleanup.test()
}
\keyword{DIANN}
