% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_go.R
\name{enrich_go}
\alias{enrich_go}
\title{Enrich GO terms for a set of genes
Runs clusterProfiler's enrichGO against a background universe, filters and simplifies
the results, and generates dotplot, cnet, and STRING network visualizations.}
\usage{
enrich_go(
  df,
  universe,
  gene_col = "Protein",
  lfc_col = "logFC",
  type = "Upregulated_all",
  go_type = "BP",
  min_count = 3
)
}
\arguments{
\item{df}{Stats output, significant proteins. One column is gene names, and one column includes the stats}

\item{universe}{Background universe to use fo enrichemnt}

\item{gene_col}{Name of column with Gene names}

\item{lfc_col}{Name of column with ranking stats}

\item{type}{String used to identify the type of comparison}

\item{go_type}{Ontology being tested. Either BP, CC, or MF}

\item{min_count}{Minimum number of proteins needed in a GO term}
}
\description{
Enrich GO terms for a set of genes
Runs clusterProfiler's enrichGO against a background universe, filters and simplifies
the results, and generates dotplot, cnet, and STRING network visualizations.
}
\examples{
enrich_go()
}
\keyword{Enrich}
\keyword{GO}
