% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_kegg.R
\name{enrich_kegg}
\alias{enrich_kegg}
\title{Enrich KEGG pathways for a set of genes}
\usage{
enrich_kegg(
  df,
  universe = universe,
  gene_col = "Protein",
  lfc_col = "logFC",
  type = paste0(unique(df$Contrast), "-all-up")
)
}
\arguments{
\item{df}{Stats output, only the significant ones. One column is gene names, and one column includes the stats}

\item{universe}{Background universe to use fo enrichemnt}

\item{gene_col}{Name of column with Gene names}

\item{lfc_col}{Name of column with ranking stats}

\item{type}{String used to identify the type of comparison}
}
\description{
Runs clusterProfiler's enrichKEGG against a background universe, clusters similar terms
by Jaccard similarity, and generates dotplot, cnet, STRING, and pathview visualizations.
}
\examples{
enrich_kegg()
}
\keyword{Enrich}
\keyword{KEGG}
