% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_kegg.R
\name{gsea_kegg}
\alias{gsea_kegg}
\title{This function performs GSEA KEGG from clsuter profiler and outputs graphs}
\usage{
gsea_kegg(
  df,
  gene_col = "Protein",
  lfc_col = "logFC",
  type = paste0(unique(df$Contrast), "-all"),
  n_perm = 10000
)
}
\arguments{
\item{df}{Stats output. One column is gene names, and one column includes the stats}

\item{gene_col}{Name of column with Gene names}

\item{lfc_col}{Name of column with ranking stats}

\item{type}{String used to identify the type of comparison}

\item{n_perm}{Number of permutations when running GSEA}
}
\description{
This function performs GSEA KEGG from clsuter profiler and outputs graphs
}
\examples{
gsea_kegg()
}
\keyword{GSEAKegg}
