% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{make_cnet_plot}
\alias{make_cnet_plot}
\title{Internal helper for enrichment network plots}
\usage{
make_cnet_plot(bio_result, go_type = "BP", type = "helper", bio_type = "GO")
}
\description{
Internal helper for enrichment network plots
}
\keyword{internal}
