% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_function.R
\name{reactome_function}
\alias{reactome_function}
\title{Reactome pathway enrichment for a set of genes
Runs ReactomePA's enrichPathway against a background universe, clusters similar terms,
and generates dotplot, cnet, STRING, and pathway view visualizations.}
\usage{
reactome_function(
  df,
  universe_df,
  type = "Upregulated",
  gene_col = "Protein",
  lfc_col = "logFC",
  top_n_pathways = 50,
  lfc_cutoff = 0.5,
  min_count = 3,
  viewpath_n = 20,
  padj_fallback = 0.2
)
}
\arguments{
\item{df}{Stats output, significant proteins. One column is gene names, and one column includes the stats}

\item{type}{String used to identify the type of comparison}

\item{gene_col}{Name of column with Gene names}

\item{lfc_col}{Name of column with ranking stats}

\item{top_n_pathways}{Number of pathways to display in dotplot}

\item{min_count}{Minimum number of proteins needed in a GO term}

\item{viewpath_n}{Number of pathways from Pathview}

\item{padj_fallback}{Max padj if nothing significant @ padj<=0.05}

\item{universe}{Background universe to use fo enrichemnt}
}
\description{
Reactome pathway enrichment for a set of genes
Runs ReactomePA's enrichPathway against a background universe, clusters similar terms,
and generates dotplot, cnet, STRING, and pathway view visualizations.
}
\examples{
reactome_function()
}
\keyword{Reactome}
