% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.testing.R
\name{statistical.testing.old}
\alias{statistical.testing.old}
\title{This function conducts t-tests across all parameters and corresponds to imputation}
\usage{
statistical.testing.old(DF, directory)
}
\arguments{
\item{DF}{output from either diann.cleanup or maxquant.cleanup}
}
\description{
This function conducts t-tests across all parameters and corresponds to imputation
}
\examples{
statistical.testing.old()
}
\keyword{stats}
