% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.testing_testing.R
\name{statistical.testing.test}
\alias{statistical.testing.test}
\title{This function conducts t-tests across all parameters}
\usage{
statistical.testing.test(DF, directory)
}
\arguments{
\item{DF}{output from either diann.cleanup or maxquant.cleanup}

\item{directory}{Output directory}
}
\description{
This function conducts t-tests across all parameters
}
\examples{
statistical.testing.test()
}
\keyword{stats}
