% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diann_cleanup.R
\name{diann.cleanup.old}
\alias{diann.cleanup.old}
\title{This function cleans up and analyzes the DIA output from DIANN using imputation}
\usage{
diann.cleanup.old(df, kinases, metadata)
}
\arguments{
\item{df}{Unique protein groups matrix from DIANN}

\item{kinases}{Either the human or mouse kinome spreadsheet}

\item{metadata}{File with column titled "Sample.ID" with the column names of the abundance values in the protein groups file, and a column titled "Treatment" with the corresponding treatment replicate}

\item{peptides}{report.parquet from DIANN}
}
\description{
This function cleans up and analyzes the DIA output from DIANN using imputation
}
\examples{
diann.cleanup.old()
}
\keyword{DIANN}
