% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diann_cleanup.R
\name{diann.cleanup}
\alias{diann.cleanup}
\title{This function cleans up and analyzes the DIA output from DIANN}
\usage{
diann.cleanup(df, kinases, peptides)
}
\arguments{
\item{df}{Unique protein groups matrix from DIANN}

\item{kinases}{Either the human or mouse kinome spreadsheet}

\item{peptides}{report.parquet from DIANN}
}
\description{
This function cleans up and analyzes the DIA output from DIANN
}
\examples{
diann.cleanup()
}
\keyword{DIANN}
