% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxquant.cleanup.R
\name{maxquant.cleanup}
\alias{maxquant.cleanup}
\title{This function cleans up and analyzes the DDA output from MaxQuant LFQ}
\usage{
maxquant.cleanup(df, kinases)
}
\arguments{
\item{df}{proteinGroups.csv from MaxQuant}

\item{kinases}{Either the human or mouse kinome spreadsheet}
}
\description{
This function cleans up and analyzes the DDA output from MaxQuant LFQ
}
\examples{
maxquant.cleanup()

}
\keyword{DIANN}
