% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.testing.R
\name{statistical.testing}
\alias{statistical.testing}
\title{This function conducts t-tests across all parameters}
\usage{
statistical.testing(DF)
}
\arguments{
\item{DF}{output from either diann.cleanup or maxquant.cleanup}
}
\description{
This function conducts t-tests across all parameters
}
\examples{
statistical.testing()
}
\keyword{stats}
